+++ Editor/SettingsInspector.cs
@@ -144,8 +144,8 @@ namespace UnityBHL
       var hotReloadProp = serializedObject.FindProperty(nameof(Settings.hotReloadOnRecompile));
       EditorGUILayout.PropertyField(hotReloadProp, new GUIContent("Hot Reload On Recompile", hotReloadProp.tooltip));
 
-      var forceOnPlayProp = serializedObject.FindProperty(nameof(Settings.forceRecompileOnPlay));
-      EditorGUILayout.PropertyField(forceOnPlayProp, new GUIContent("Recompile On Play", forceOnPlayProp.tooltip));
+      var recompileOnPlayProp = serializedObject.FindProperty(nameof(Settings.recompileOnPlay));
+      EditorGUILayout.PropertyField(recompileOnPlayProp, new GUIContent("Recompile On Play", recompileOnPlayProp.tooltip));
 
       serializedObject.ApplyModifiedProperties();
     }